Restructure in prep for flask-restful
