+++ common/functions.py
